backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-dotenv==1.0.1
pyyaml==6.0.2
PyGithub==2.4.0
httpx==0.27.2
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-openai==0.2.9
langgraph==0.2.38
python-multipart==0.0.12
pydantic==2.9.2
aiohttp==3.10.10
